Correct icon button border in dark theme
When `IconButton` is used in dark theme, the static border is set to 0,
while in light theme it is set to 0.03125rem.

On hover, in dark theme, the border is set to 0.03125rem which results
into a jumpy behavior as the dimension changes from 0 to 0.03125rem.

This change sets the border color to be transparent and matches
the border dimension with light theme.