Add ability to override event reporter when using React library
Add package.json to serverless src

This helps us in installing updated dependency like aws-sdk
which the lambda environment does not update. This is required
since event ingestion URL is only released in new aws-sdk version
and we need that version to test and enable event reporting.

Add changelog entry