Add `MeetingStatus.Reconnecting` and `audioVideoDidStartConnecting` observer
- `audioVideoDidStartConnecting` helps in listening to whether JS SDK is
reconnecting.

- Applications depending on Chime SDK component library currently do not
have a out of the box `MeetingStatus` to tell whether the JS SDK is in reconnection.
Thus, add `audioVideoDidStartConnecting` with a new `MeetingStatus.Reconnecting`
to aid such a use case.

- Fix: Publish `MeetingStatus.Failed` only when a non-terminal failure
is encountered. Currently, some statuses pass both the `sessionStatus.isFailure()`
as well as `sessionStatus.isTerminal()` in JS SDK, thus, for a status if it is in both,
then it will be considered as failure and `MeetingStatus.TerminalFailure` will never
get set. Check below file for more information on both methods in JS SDK:
https://github.com/aws/amazon-chime-sdk-js/blob/main/src/meetingsession/MeetingSessionStatus.ts